+++ changelog.md
@@ -9,7 +9,7 @@
 - Unlink your Faction so that you can link it with another
 
 ### Bug Fixes
-- Ability bar height bug found by Blaise
+- Paragraph height bug found by Blaise
 
 ## [1.12.3] - 2026-5-13 Forging Cards